+++ shared/r2-shared-swift/Fetcher/Resource/DataResource.swift
@@ -32,6 +32,8 @@ public final class DataResource: Resource {
     
     public let link: Link
     
+    public let file: URL? = nil
+    
     public var length: ResourceResult<UInt64> { data.map { UInt64($0.count) } }
     
     public func read(range: Range<UInt64>?) -> ResourceResult<Data> {