+++ shared/r2-shared-swift/Toolkit/PDF/PDFKit.swift
@@ -45,6 +45,10 @@ public class PDFKitPDFDocumentFactory: PDFDocumentFactory {
     }
     
     public func open(resource: Resource, password: String?) throws -> PDFDocument {
+        if let url = resource.file {
+            return try open(url: url, password: password)
+        }
+        
         // Unfortunately, PDFKit doesn't support streams, so we need to load the full document in
         // memory. If this is an issue for you, use `CPDFDocumentFactory` instead.
         guard